fjerner mer som ikke er i bruk, og flytter litt på noe
