+++ src/Forvaltningsportalen/RightWindow.js
@@ -1,4 +1,4 @@
-import ForvaltningsKartlag from "Forvaltningsportalen/ForvaltningsKartlag/ForvaltningsKartlag";
+import ForvaltningsKartlag from "./ForvaltningsKartlag/ForvaltningsKartlag";
 import React from "react";
 
 const RightWindow = props => {
